+++ scripts/valistRelease.js
@@ -23,7 +23,7 @@ const download = async ({ url, path }) => {
 
 async function publishToValist(releaseDetails) {
 
-    const web3 = new Web3HttpProvider('https://polygon-rpc.com/');
+    const web3 = new Web3HttpProvider('https://polygon-rpc.com');
     const provider = new ethers.providers.Web3Provider(web3);
 
     const wallet = new ethers.Wallet(process.env.VALIST_RELEASER_PK, provider);